requirements.txt
aiofile
aiopg
asyncpg
databases[postgresql]
fastapi
httpx
psycopg2
python-jose[cryptography]
python-multipart
pyyaml
requests
SQLAlchemy
uvicorn
